+++ CHANGELOG.md
@@ -6,7 +6,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [4.64]
 ### Added
-- Create/Update a gene panel contaning all PanelApp green genes (`scout update panelapp-green -i <cust_id>`)
+- Create/Update a gene panel containing all PanelApp green genes (`scout update panelapp-green -i <cust_id>`)
 ### Changed
 - Open local observation matching cases in new windows
 ### Fixed